Add basic journal entry analysis functionality
Broke out a couple of commands that work together to build up a prompt
containing some instructions and a journal entry to give to OpenAI for
analysis text generation.

Add a new `analysis` attribute to the JournalEntry model that is
populated with text generated by the language model.

Analysis happens on the journal entry show, so the form that was being
rendered was replace with read-only title and content. The "edit"
template was updated to render the proper form and links were added to
show and index to access the edit action.