Try out a mobile layout for smaller screens
There are two bits of javascript that are hacks to get the appearance
just right in mobile browsers where the address bar might change the
document flow in undesired ways.

The `resizeMain` function makes it possible for the content to take up
exactly the viewable area. This is a workaround that allows the footer
to be visible without having the scroll the address bar away on android
and other mobile browsers.

The `adjustScrollableHeight` function forces the height of the main
scrollable area on `div#scrollable` elements to fit to the screen. On
mobile browsers, the address bar will sometimes push the bottom of the
content off the viewable area, preventing the scroll from revealing all
the content.