Add basic core values functionality
This commit adds the ability for the user to enter core values. The
intention is to eventually hook this up to the reflection tools to
better help the user focus on what's important to them.

This commit also tweaks the submit_state_controller functionality so
that forms can't be submitted if certain fields are marked as required