Use requestSubmit instead of submit
There are a couple examples of stimulus controllers in this app that are
modifying a submit event for the targeted form. There seems to be an
issue with Turbo where using `submit` doesn't create a request using the
turbo stream format. For whatever reason, the `requestSubmit` function
does trigger the form submit with the proper format.