Use `update` method instead of `replace` for chat log entries
This commit swaps out a turbostream method for update. Previously the
`replace` method was being used and that was also replacing the target
div. This caused any subsequent requests to break because there was no
div to apply the new turbo content to. Using `update` replaces the inner
contents of the target leaving the target itself intact.