Drop feather-icons package for custom solution
This commit drops the feather-icons package in favor of a simple set of
helper methods for rendering SVGs inline. Included in this commit is the
complete library of Feather icons downloaded from the [Feather
site](https://feathericons.com/).

The main reason for this change was to drop the need to call
`feather.replace()` anytime an icon needed to be loaded. Normally this
only has to happen on initial page load, but because of the extensive
use of Turbo, this doesn't catch every case. It was annoying to have to
keep track of every place where `replace` was needed or come up with a
solution that plugged that into turbo.

This solution is much simpler and will probably work for now.