Refactor `StreakCounter#longest_streak` to use SQL query
This commit updates the longest_streak method so that the calculation is
no longer being done by iterating over records in memory.

The new approach uses the `RANK()` SQL function to help generate
grouping identifiers and group contiguous dates together. Once those are
dates are group, counts for number of records in each group are
calculated. Finally, the max of those counts is found. This number
represents the length of the longest streak.

This new approach was written in a way that will work for all
ActiveRecord classes with a `created_at` column.