Introduce a User model
This commit introduces a `User` model so that all data stored in the app
can be stored and access by only the user who created it.

The `clearance` gem was added to the application to provide support for
email and password authentication by users. This replaces the HTTP basic
auth that was protecting the app before.

Controllers were updated to inherit from a new `AuthenticatedController`
that requires log in for all current actions in the application. Most
actions now access resources in the app via the `current_user` variable
provided by `clearance`.

Future work should introduce additional means for ensuring that the data
being loaded in a given action is appropriate for the current user.
Pundit might be worth introducing to provide authorization tooling.